Moved all my line-level parsers (and the one for comment refs) into a lower namespace to keep the Parsing namespace tidy (full of entry points).
