+++ src/Core/Gibberish/Gibberish.csproj
@@ -101,13 +101,13 @@
     <Compile Include="AST\_1_Bare\LanguageConstructVisitor.cs" />
     <Compile Include="AST\ParseError.cs" />
     <Compile Include="AST\_1_Bare\CommentDefinitionBlock.cs" />
-    <Compile Include="Parsing\ParseBlankLines.cs" />
-    <Compile Include="Parsing\ParseCommentDefinitions.cs" />
-    <Compile Include="Parsing\ParseCommentDefinitionPreludes.cs" />
-    <Compile Include="Parsing\ParseCommentReferences.cs" />
-    <Compile Include="Parsing\ParseCommentStatement.cs" />
-    <Compile Include="Parsing\ParsePreludes.cs" />
-    <Compile Include="Parsing\ParseStatements.cs" />
+    <Compile Include="Parsing\LineParsers\ParseBlankLines.cs" />
+    <Compile Include="Parsing\LineParsers\ParseCommentDefinitions.cs" />
+    <Compile Include="Parsing\LineParsers\ParseCommentDefinitionPreludes.cs" />
+    <Compile Include="Parsing\LineParsers\ParseCommentReferences.cs" />
+    <Compile Include="Parsing\LineParsers\ParseCommentStatements.cs" />
+    <Compile Include="Parsing\LineParsers\ParsePreludes.cs" />
+    <Compile Include="Parsing\LineParsers\ParseStatements.cs" />
     <Compile Include="Parsing\ParsingExtensions.cs" />
     <Compile Include="AST\_1_Bare\UnknownPrelude.cs" />
     <Compile Include="Parsing\RecognizeLines.cs" />
