+++ src/Core/Gibberish/Parsing/LineParsers/ParseCommentDefinitions.cs
@@ -4,7 +4,7 @@ using Gibberish.AST;
 using Gibberish.AST._1_Bare;
 using JetBrains.Annotations;
 
-namespace Gibberish.Parsing
+namespace Gibberish.Parsing.LineParsers
 {
 	public static class ParseCommentDefinitions
 	{
